added some more questions to the script

+++ neon-init.sh
@@ -1,25 +1,24 @@
 #!/bin/sh
 
 # make "native" folder
-if [ -d native ]
+echo -n "Do you want to enable create 'native' folder? (y/n)? "
+read answer
+if echo "$answer" | grep -iq "^y" ;
 then
-    echo "'native' folder exist"
-    exit 1
-fi
+    if [ -d native ]
+    then
+        echo "'native' folder exist"
+        exit 1
+    fi
 
-mkdir -p native/src
+    mkdir -p native/src
 
-# download template files
-curl -o native/Cargo.toml https://raw.githubusercontent.com/benjcal/neon-init/master/templates/Cargo.toml
-curl -o native/build.rs https://raw.githubusercontent.com/benjcal/neon-init/master/templates/build.rs
-curl -o native/src/lib.rs https://raw.githubusercontent.com/benjcal/neon-init/master/templates/src/lib.rs
-
-
-npm install neon-cli
-
-if [ -f package.json ]
-then
-    echo "package.json not found"
+    # download template files
+    curl -o native/Cargo.toml https://raw.githubusercontent.com/benjcal/neon-init/master/templates/Cargo.toml
+    curl -o native/build.rs https://raw.githubusercontent.com/benjcal/neon-init/master/templates/build.rs
+    curl -o native/src/lib.rs https://raw.githubusercontent.com/benjcal/neon-init/master/templates/src/lib.rs
+else
+    echo "can't install neon without a native folder"
     exit 1
 fi
 
@@ -27,6 +26,16 @@ echo -n "Do you want to enable 'npm run build' for this project? (y/n)? "
 read answer
 if echo "$answer" | grep -iq "^y" ;
 then
+
+    if [ -f package.json ]
+    then
+        echo "package.json not found"
+        echo "generate one using 'npm init'"
+        exit 1
+    fi
+
+    npm install neon-cli
+
     # add neon-build to package.json
     node -e "const fs = require('fs'); \
     let p = fs.readFileSync('package.json', 'utf8'); \
@@ -38,5 +47,3 @@ else
     echo "you can build the neon library by running"
     echo "./node_modules/neon-cli/bin/cli.js build"
 fi
-
-
